+++ strategy/__init__.py
@@ -12,6 +12,7 @@ __all__ = [
     "event_confluence",
     "event_context",
     "event_diagnostics",
+    "event_research_summary",
     "filters",
     "presets",
     "preset_comparison",